Rename subcommand from ecguess to guess
- Update internal/cli/cli.go to use `guess` as subcommand name.
- Regenerate CLI code, which moved `cmd/ecguess` to `cmd/guess`.
- Apply manual fixes to `cmd/guess/root.go` (missing import, qualifier).
- Ensure generated code is lint-free.
