--- cmd/ecguess/main.go
@@ -1,29 +1,0 @@
-// Generated by github.com/arran4/go-subcommand/cmd/gosubc
-
-package main
-
-//go:generate sh -c "command -v gosubc >/dev/null 2>&1 && gosubc generate . || go run github.com/arran4/go-subcommand/cmd/gosubc generate ."
-
-import (
-	"fmt"
-	"os"
-)
-
-var (
-	version = "dev"
-	commit  = "none"
-	date    = "unknown"
-)
-
-func main() {
-	root, err := NewRoot("ecguess", version, commit, date)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
-		os.Exit(1)
-	}
-
-	if err := root.Execute(os.Args[1:]); err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %v\n", err)
-		os.Exit(1)
-	}
-}
