--- cmd/ecguess/templates/dummy.txt
@@ -1,1 +1,0 @@
-Dummy
