+++ cmd/guess/root_test.go
@@ -1,0 +1,31 @@
+// Generated by github.com/arran4/go-subcommand/cmd/gosubc
+
+package main
+
+import (
+	"testing"
+)
+
+func TestRoot_Execute(t *testing.T) {
+	cmd, err := NewRoot("test", "", "", "")
+	if err != nil {
+		t.Fatalf("Failed to create root command: %v", err)
+	}
+
+	called := false
+	cmd.CommandAction = func(c *RootCmd) error {
+		called = true
+		return nil
+	}
+
+	args := []string{}
+
+	err = cmd.Execute(args)
+	if err != nil {
+		t.Errorf("Unexpected error: %v", err)
+	}
+	if !called {
+		t.Error("CommandAction was not called")
+	}
+
+}
