+++ cmd/guess/templates/dummy.txt
@@ -1,0 +1,1 @@
+Dummy
